# Flatten nested string concat