@kaikalur 2023-10-10
Performance benchmark results:

```

New  version with concat flattened
sql_concat_1k ::  256.421 cpu ms :: 6.6MB peak memory :: in    25,      0B,      97/s,      0B/s :: out  1000,  57.2KB,    3.9K/s,   223KB/s}

Original:
sql_concat_1k ::  260.717 cpu ms :: 6.6MB peak memory :: in    25,      0B,      95/s,      0B/s :: out  1000,  57.2KB,   3.83K/s,   219KB/s
```